+++ src/components/Controls.vue
@@ -34,7 +34,7 @@
 			<h2><a href="#">{{ overviewName }}</a></h2>
 			<Actions>
 				<ActionButton icon="icon-add" @click.stop="modalShow=true">
-					{{ t('deck', 'Add card on Today') }}
+					{{ t('deck', 'Add card on today') }}
 				</ActionButton>
 			</Actions>
 		</div>
@@ -199,55 +199,20 @@
 			</div>
 		</div>
 
-		<Modal v-if="modalShow" :title="t('deck', 'Add card on Today')" @close="modalShow=false">
-			<div class="modal__content">
-				{{ lastBoardId }}
-				{{ lastListId }}
-				<h3>{{ t('deck', 'Add card on Today') }}</h3>
-				<Multiselect v-model="selectedBoard"
-					:placeholder="t('deck', 'Select a board')"
-					:options="boards"
-					:max-height="100"
-					label="title"
-					@select="loadStacksFromBoard" />
-				<Multiselect v-model="selectedStack"
-					:placeholder="t('deck', 'Select a list')"
-					:options="stacksFromBoard"
-					:max-height="100"
-					label="title" />
-
-				<label for="new-stack-input-main" class="hidden-visually">{{ t('deck', 'Add card on Today') }}</label>
-				<input id="new-stack-input-main"
-					ref="newCardInput"
-					v-model="newCardTitle"
-					v-focus
-					type="text"
-					class="no-close"
-					:placeholder="t('deck', 'Card name')">
-
-				<button :disabled="!addCardSetRequiredFields" class="primary" @click="addCard">
-					{{ t('deck', 'Add card') }}
-				</button>
-				<button @click="modalShow=false">
-					{{ t('deck', 'Cancel') }}
-				</button>
-			</div>
-		</Modal>
+		<ControlsModal :modal-show="modalShow" />
 	</div>
 </template>
 
 <script>
 import { mapState, mapGetters } from 'vuex'
-import { Modal, Multiselect, Actions, ActionButton, Popover, Avatar } from '@nextcloud/vue'
+import { Actions, ActionButton, Popover, Avatar } from '@nextcloud/vue'
 import labelStyle from '../mixins/labelStyle'
-import axios from '@nextcloud/axios'
-import { generateUrl } from '@nextcloud/router'
-import { showError } from '@nextcloud/dialogs'
+import ControlsModal from './ControlsModal'
 
 export default {
 	name: 'Controls',
 	components: {
-		Actions, ActionButton, Popover, Avatar, Modal, Multiselect,
+		Actions, ActionButton, Popover, Avatar, ControlsModal,
 	},
 	mixins: [ labelStyle ],
 	props: {
@@ -322,15 +287,6 @@ export default {
 		board() {
 			this.clearFilter()
 		},
-		lastBoardId() {
-
-			if (this.lastBoardId === null || this.lastBoardId === 0) {
-				return
-			}
-			this.selectedBoard = this.boards.filter(board => {
-				return board.id === this.lastBoardId
-			})
-		},
 	},
 	methods: {
 		beforeSetFilter(e) {
@@ -383,33 +339,6 @@ export default {
 			this.$store.dispatch('setFilter', { ...filterReset })
 			this.filter = filterReset
 		},
-		async loadStacksFromBoard(selectedBoard) {
-			try {
-				const url = generateUrl('/apps/deck/stacks/' + selectedBoard.id)
-				const response = await axios.get(url)
-				this.stacksFromBoard = response.data
-			} catch (err) {
-				return err
-			}
-		},
-		async addCard() {
-			try {
-				const today = new Date()
-				today.setHours(23, 59, 59, 999)
-				await this.$store.dispatch('addCard', {
-					title: this.newCardTitle,
-					stackId: this.selectedStack.id,
-					boardId: this.selectedBoard.id,
-					duedate: today.toISOString(),
-				})
-				this.newCardTitle = ''
-				this.$store.dispatch('storeLastListId', this.selectedStack.id)
-				this.$store.dispatch('storeLastBoardId', this.selectedBoard.id)
-			} catch (e) {
-				showError('Could not create card: ' + e.response.data.message)
-			}
-			this.modalShow = false
-		},
 	},
 }
 </script>
@@ -417,23 +346,6 @@ export default {
 <style lang="scss" scoped>
 	#new-stack-input-main {
 		width: 100%;
-	}
-
-	.modal__content {
-		width: 25vw;
-		min-width: 250px;
-		min-height: 120px;
-		text-align: center;
-		margin: 20px 20px 100px 20px;
-
-		.multiselect {
-			margin-bottom: 10px;
-		}
-	}
-
-	.modal__content button {
-		float: right;
-		margin-top: 50px;
 	}
 
 	.controls {
